feat(#13): separate tests module
