+++ tests/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "tests"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.86"
+
+[dev-dependencies]
+assert_cmd = "2.0.14"
+cli = { path = "../cli" }
